Move region definition to separate file
